+++ custom_components/ev_savings/const.py
@@ -1,5 +1,5 @@
 """Shared constants."""
 DOMAIN = "ev_savings"
-VERSION = "0.6.2"
+VERSION = "0.6.3"
 ROLES = ("discount", "off_peak", "peak")
 DUKE_PROFILE = "duke_fl_rst1"